Remove spaceport (pre test)
