+++ apollo-router/Cargo.toml
@@ -61,7 +61,6 @@ clap = { version = "3.2.23", default-features = false, features = [
 console-subscriber = { version = "0.1.8", optional = true }
 ctor = "0.1.23"
 dashmap = { version = "5.4.0", features = ["serde"] }
-deadpool = { version = "0.9.5", features = ["rt_tokio_1"] }
 derivative = "2.2.0"
 derive_more = { version = "0.99.17", default-features = false, features = [
     "from",
