+++ apollo-router/src/plugins/telemetry/apollo_exporter.rs
@@ -1,31 +1,29 @@
 //! Configuration for apollo telemetry exporter.
 // This entire file is license key functionality
-use std::time::Duration;
-
-use async_trait::async_trait;
-use deadpool::managed;
-use deadpool::managed::Pool;
-use deadpool::managed::RecycleError;
-use deadpool::Runtime;
+use bytes::BytesMut;
+use flate2::write::GzEncoder;
+use flate2::Compression;
 use futures::channel::mpsc;
 use futures::stream::StreamExt;
+use http::header::CONTENT_TYPE;
+use reqwest::Client;
+use std::io::Write;
+use std::time::Duration;
 use sys_info::hostname;
+use tonic::Status;
 use tower::BoxError;
 use url::Url;
 
 use super::apollo::Report;
 use super::apollo::SingleReport;
-use crate::spaceport::ReportHeader;
-use crate::spaceport::Reporter;
-use crate::spaceport::ReporterError;
-// use crate::plugins::telemetry::apollo::ReportBuilder;
 
 const DEFAULT_QUEUE_SIZE: usize = 65_536;
-const DEADPOOL_SIZE: usize = 128;
 // Do not set to 5 secs because it's also the default value for the BatchSpanProcesser of tracing.
 // It's less error prone to set a different value to let us compute traces and metrics
 pub(crate) const EXPORTER_TIMEOUT_DURATION: Duration = Duration::from_secs(6);
-pub(crate) const POOL_TIMEOUT: Duration = Duration::from_secs(5);
+static DEFAULT_APOLLO_USAGE_REPORTING_INGRESS_URL: &str =
+    "https://usage-reporting.api.apollographql.com/api/ingress/traces";
+const BACKOFF_INCREMENT: Duration = Duration::from_millis(50);
 
 #[derive(Clone)]
 pub(crate) enum Sender {
@@ -74,7 +72,7 @@ impl ApolloExporter {
         // * If the exporter is dropped the remaining records are flushed.
         let (tx, mut rx) = mpsc::channel::<SingleReport>(DEFAULT_QUEUE_SIZE);
 
-        let header = crate::spaceport::ReportHeader {
+        let header = crate::plugins::telemetry::apollo_exporter::proto::ReportHeader {
             graph_ref: apollo_graph_ref.to_string(),
             hostname: hostname()?,
             agent_version: format!(
@@ -88,37 +86,12 @@ impl ApolloExporter {
             ..Default::default()
         };
 
-        // Pool Sizing: by default Deadpool will configure a maximum
-        // pool size based on the number of physical CPUs:
-        //   `cpu_count * 4` ignoring any logical CPUs (Hyper-Threading).
-        // This is going to be very low in containerised environments
-        // For example, in my k8s testing I get max_size: 16
-        //
-        // Since we know we can support large numbers of connections to the
-        // data ingestion endpoint, I'm going to manually set this to
-        // be [`DEADPOOL_SIZE`] which should be plenty. I'm not setting
-        // it to be a very high number (e.g.: 100000), because resources
-        // are consumed and conserving them is important.
-
-        //
-        // Deadpool gives us connection pooling to spaceport
-        // It also significantly simplifies initialisation of the connection and gives us options in the future for configuring timeouts.
-        let pool = deadpool::managed::Pool::<ReporterManager>::builder(ReporterManager {
-            endpoint: endpoint.clone(),
-        })
-        .max_size(DEADPOOL_SIZE)
-        .create_timeout(Some(POOL_TIMEOUT))
-        .recycle_timeout(Some(POOL_TIMEOUT))
-        .wait_timeout(Some(POOL_TIMEOUT))
-        .runtime(Runtime::Tokio1)
-        .build()
-        .unwrap();
-
         let spaceport_endpoint = endpoint.clone();
-        tracing::info!(%spaceport_endpoint, "creating apollo exporter");
+        tracing::debug!(%spaceport_endpoint, "creating apollo exporter");
 
         // This is the task that actually sends metrics
         tokio::spawn(async move {
+            let client = reqwest::Client::new();
             let timeout = tokio::time::interval(EXPORTER_TIMEOUT_DURATION);
             let mut report = Report::default();
 
@@ -130,17 +103,27 @@ impl ApolloExporter {
                         if let Some(r) = single_report {
                             report += r;
                         } else {
-                            tracing::info!(%spaceport_endpoint, "terminating apollo exporter");
+                            tracing::debug!(%spaceport_endpoint, "terminating apollo exporter");
                             break;
                         }
                        },
                     _ = timeout.tick() => {
-                        Self::send_report(&pool, &apollo_key, &header, std::mem::take(&mut report)).await;
+                        if let Err(e) = Self::submit_report(&client, &apollo_key, std::mem::take(&mut report).into_report(header.clone())).await {
+                            tracing::error!("Failed to submit Apollo report: {}", e)
+                        }
                     }
                 };
             }
 
-            Self::send_report(&pool, &apollo_key, &header, report).await;
+            if let Err(e) = Self::submit_report(
+                &client,
+                &apollo_key,
+                std::mem::take(&mut report).into_report(header),
+            )
+            .await
+            {
+                tracing::error!("Failed to submit Apollo report: {}", e)
+            }
         });
         Ok(ApolloExporter { tx })
     }
@@ -149,61 +132,87 @@ impl ApolloExporter {
         Sender::Spaceport(self.tx.clone())
     }
 
-    async fn send_report(
-        pool: &Pool<ReporterManager>,
-        apollo_key: &str,
-        header: &ReportHeader,
-        report: Report,
-    ) {
-        if report.operation_count == 0 && report.traces_per_query.is_empty() {
-            return;
-        }
-
-        match pool.get().await {
-            Ok(mut reporter) => {
-                let report = report.into_report(header.clone());
-                match reporter
-                    .submit(crate::spaceport::ReporterRequest {
-                        apollo_key: apollo_key.to_string(),
-                        report: Some(report),
-                    })
-                    .await
-                {
-                    Ok(_) => {}
-                    Err(e) => {
-                        tracing::warn!("failed to submit stats to spaceport: {}", e);
-                    }
-                };
-            }
-            Err(err) => {
-                tracing::warn!(
-                    "stats discarded as unable to get connection to spaceport: {}",
-                    err
-                );
-            }
+    pub(crate) async fn submit_report(
+        client: &Client,
+        key: &str,
+        report: crate::plugins::telemetry::apollo_exporter::proto::Report,
+    ) -> Result<(), Status> {
+        tracing::debug!("submitting report: {:?}", report);
+        // Protobuf encode message
+        let mut content = BytesMut::new();
+        prost::Message::encode(&report, &mut content)
+            .map_err(|e| Status::invalid_argument(e.to_string()))?;
+        // Create a gzip encoder
+        let mut encoder = GzEncoder::new(Vec::new(), Compression::default());
+        // Write our content to our encoder
+        encoder
+            .write_all(&content)
+            .map_err(|e| Status::internal(e.to_string()))?;
+        // Finish encoding and retrieve content
+        let compressed_content = encoder
+            .finish()
+            .map_err(|e| Status::internal(e.to_string()))?;
+        let mut backoff = Duration::from_millis(0);
+        let ingress = match std::env::var("APOLLO_USAGE_REPORTING_INGRESS_URL") {
+            Ok(v) => v,
+            Err(_e) => DEFAULT_APOLLO_USAGE_REPORTING_INGRESS_URL.to_string(),
         };
-    }
-}
+        let req = client
+            .post(ingress)
+            .body(compressed_content)
+            .header("X-Api-Key", key)
+            .header("Content-Encoding", "gzip")
+            .header(CONTENT_TYPE, "application/protobuf")
+            .header("Accept", "application/json")
+            .header(
+                "User-Agent",
+                format!(
+                    "{} / {} usage reporting",
+                    std::env!("CARGO_PKG_NAME"),
+                    std::env!("CARGO_PKG_VERSION")
+                ),
+            )
+            .build()
+            .map_err(|e| Status::unavailable(e.to_string()))?;
 
-#[derive(Debug)]
-pub(crate) struct ReporterManager {
-    endpoint: Url,
-}
-
-#[async_trait]
-impl managed::Manager for ReporterManager {
-    type Type = Reporter;
-    type Error = ReporterError;
-
-    async fn create(&self) -> Result<Reporter, Self::Error> {
-        tracing::debug!("creating reporter: {:?}", self.endpoint);
-        let url = self.endpoint.to_string();
-        Reporter::try_new(url).await
-    }
-
-    async fn recycle(&self, r: &mut Reporter) -> managed::RecycleResult<Self::Error> {
-        tracing::debug!("recycling reporter: {:?}", r);
-        r.reconnect().await.map_err(RecycleError::Backend)
+        let mut msg = "default error message".to_string();
+        for i in 0..5 {
+            // We know these requests can be cloned
+            let task_req = req.try_clone().expect("requests must be clone-able");
+            match client.execute(task_req).await {
+                Ok(v) => {
+                    let status = v.status();
+                    let data = v
+                        .text()
+                        .await
+                        .map_err(|e| Status::internal(e.to_string()))?;
+                    // Handle various kinds of status:
+                    //  - if client error, terminate immediately
+                    //  - if server error, it may be transient so treat as retry-able
+                    //  - if ok, return ok
+                    if status.is_client_error() {
+                        tracing::error!("client error reported at ingress: {}", data);
+                        return Err(Status::invalid_argument(data));
+                    } else if status.is_server_error() {
+                        tracing::warn!("attempt: {}, could not transfer: {}", i + 1, data);
+                        msg = data;
+                    } else {
+                        tracing::debug!("ingress response text: {:?}", data);
+                        return Ok(());
+                    }
+                }
+                Err(e) => {
+                    // TODO: Ultimately need more sophisticated handling here. For example
+                    // a redirect should not be treated the same way as a connect or a
+                    // type builder error...
+                    tracing::warn!("attempt: {}, could not transfer: {}", i + 1, e);
+                    msg = e.to_string();
+                }
+            }
+            backoff += BACKOFF_INCREMENT;
+            tokio::time::sleep(backoff).await;
+        }
+        Err(Status::unavailable(msg))
     }
 }
 
@@ -228,4 +237,111 @@ pub(crate) fn get_uname() -> Result<String, std::io::Error> {
         "{}, {}, {}, {}, {}",
         sysname, nodename, release, version, machine
     ))
+}
+
+#[allow(unreachable_pub)]
+pub(crate) mod proto {
+    tonic::include_proto!("report");
+}
+
+pub(crate) use prost::*;
+use serde::ser::SerializeStruct;
+use std::error::Error;
+use tokio::task::JoinError;
+use tonic::codegen::http::uri::InvalidUri;
+
+/// Reporting Error type
+#[derive(Debug)]
+pub(crate) struct ReporterError {
+    source: Box<dyn Error + Send + Sync + 'static>,
+    msg: String,
+}
+
+impl std::error::Error for ReporterError {}
+
+impl From<InvalidUri> for ReporterError {
+    fn from(error: InvalidUri) -> Self {
+        ReporterError {
+            msg: error.to_string(),
+            source: Box::new(error),
+        }
+    }
+}
+
+impl From<tonic::transport::Error> for ReporterError {
+    fn from(error: tonic::transport::Error) -> Self {
+        ReporterError {
+            msg: error.to_string(),
+            source: Box::new(error),
+        }
+    }
+}
+
+impl From<std::io::Error> for ReporterError {
+    fn from(error: std::io::Error) -> Self {
+        ReporterError {
+            msg: error.to_string(),
+            source: Box::new(error),
+        }
+    }
+}
+
+impl From<sys_info::Error> for ReporterError {
+    fn from(error: sys_info::Error) -> Self {
+        ReporterError {
+            msg: error.to_string(),
+            source: Box::new(error),
+        }
+    }
+}
+
+impl From<JoinError> for ReporterError {
+    fn from(error: JoinError) -> Self {
+        ReporterError {
+            msg: error.to_string(),
+            source: Box::new(error),
+        }
+    }
+}
+
+impl std::fmt::Display for ReporterError {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(
+            f,
+            "ReporterError: source: {}, message: {}",
+            self.source, self.msg
+        )
+    }
+}
+
+pub(crate) fn serialize_timestamp<S>(
+    timestamp: &Option<prost_types::Timestamp>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: serde::Serializer,
+{
+    match timestamp {
+        Some(ts) => {
+            let mut ts_strukt = serializer.serialize_struct("Timestamp", 2)?;
+            ts_strukt.serialize_field("seconds", &ts.seconds)?;
+            ts_strukt.serialize_field("nanos", &ts.nanos)?;
+            ts_strukt.end()
+        }
+        None => serializer.serialize_none(),
+    }
+}
+
+#[cfg(not(windows))] // git checkout converts \n to \r\n, making == below fail
+#[test]
+fn check_reports_proto_is_up_to_date() {
+    let proto_url = "https://usage-reporting.api.apollographql.com/proto/reports.proto";
+    let response = reqwest::blocking::get(proto_url).unwrap();
+    let content = response.text().unwrap();
+    // Not using assert_eq! as printing the entire file would be too verbose
+    assert!(
+        content == include_str!("proto/reports.proto"),
+        "Protobuf file is out of date. Run this command to update it:\n\n    \
+            curl -f {proto_url} > apollo-router/src/spaceport/proto/reports.proto\n\n"
+    );
 }
