--- apollo-router/src/spaceport/proto/agents.proto
@@ -1,21 +1,0 @@
-// With regards to ELv2 licensing, this entire file is license key functionality
-syntax = "proto3";
-
-import "reports.proto";
-
-package Agent;
-
-service Reporter {
-    rpc Add(ReporterRequest) returns (ReporterResponse) {}
-}
-
-message ReporterResponse {
-    string message = 1;
-}
-// This is just a wrapper around Report that supplies the Apollo key.
-// Spaceport will use this when sending data to Apollo.
-message ReporterRequest {
-    string apollo_key = 1;
-    Report.Report report = 2;
-}
-
