+++ examples/telemetry/README.md
@@ -5,7 +5,6 @@ Demonstrates configuring of the router for:
 * OpenTelemetry
   * Jaeger
   * OpenTelemetry Collector
-* Spaceport (Apollo Studio)
 
 
 ## OpenTelemetry
